05b-Agents/requirements.txt
strands-agents
strands-agents-tools
bedrock-agentcore
bedrock-agentcore-starter-toolkit
boto3
botocore
awscli
opensearch-py
requests-aws4auth
pyyaml
retrying
pandas
uv
openpyxl